feat: configurable diagonal movement (II)
* Reviewed and applied the maintainer's suggested changes.
* Massive refactoring: uses Directions.cs enums instead of integers within the solution for better code readability and consistency.
* Clean up and fixes: fixed 4 directions movement, replaced the CheckAdjacentTiles() method with a configurable server side boolean that allows it to limit auto-turn to target feature: same results and performs better.
* Critter: added ability to move diagonally when enabled.
* Dash & knock-back: added ability to dash / push diagonally when enabled.
* Changed Directions.cs order for diagonal enums (UL, UR, DR, DL).
* Replaces method GetSpriteDirection() by a readonly array field.
* Changed diagonal sprite directions: entities face up/down instead of sides when moving diagonally ("TLOZ: A Link To the Past" like style).

+++ Intersect.Client/Networking/PacketSender.cs
@@ -6,7 +6,6 @@ using Intersect.Client.General;
 using Intersect.Client.Interface.Game;
 using Intersect.Client.Maps;
 using Intersect.Network.Packets.Client;
-using Intersect.Utilities;
 
 namespace Intersect.Client.Networking
 {
@@ -41,7 +40,7 @@ namespace Intersect.Client.Networking
 
         public static void SendMove()
         {
-            Network.SendPacket(new MovePacket(Globals.Me.MapId, Globals.Me.X, Globals.Me.Y, Globals.Me.Dir));
+            Network.SendPacket(new MovePacket(Globals.Me.MapId, Globals.Me.X, Globals.Me.Y, (byte)Globals.Me.Dir));
         }
 
         public static void SendChatMsg(string msg, byte channel)
